+++ packages/core/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'core',
-  version: '0.4.7',
+  version: '0.4.8',
   source: '@nuxt-customer-portal/core',
   dependsOn: [],
   schema: './server/db/schema/index.ts',