+++ packages/invoices/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'invoices',
-  version: '0.4.7',
+  version: '0.4.8',
   source: '@nuxt-customer-portal/invoices',
   clientModuleId: 'invoices',
   dependsOn: ['core', 'ui', 'organizations', 'clients'],