+++ packages/preset/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'preset',
-  version: '0.4.7',
+  version: '0.4.8',
   source: '@nuxt-customer-portal/preset',
   dependsOn: ['core', 'ui', 'authentication', 'organizations', 'administration', 'clients'],
   includes: [