+++ packages/saas-configuration/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'saas-configuration',
-  version: '0.4.7',
+  version: '0.4.8',
   source: '@nuxt-customer-portal/saas-configuration',
   dependsOn: ['core'],
   migrations: './migrations'