+++ packages/invoice-timesheets/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nuxt-customer-portal/invoice-timesheets
 
+## 0.4.8
+
+### Patch Changes
+
+- @nuxt-customer-portal/core@0.4.8
+- @nuxt-customer-portal/invoices@0.4.8
+- @nuxt-customer-portal/timesheets@0.4.8
+
 ## 0.4.7
 
 ### Patch Changes